frontend: Consider settings changed if an output setting was reset
If the previous selection is no longer available, the UI would default
to the first index, but not allow the user to save the change (which
then becomes impossible to save if the setting only has one possible
value, since you can never trigger the "changed" status).

If the settings load process clamped any setting to the default, enable
the apply button and consider the setting changed.